move components into components folder
