+++ frontend/src/components/Note.jsx
